+++ src/main/java/org/opentmf/bpmn/sync/service/impl/ReactiveBpmnSyncServiceImpl.java
@@ -83,9 +83,10 @@ public class ReactiveBpmnSyncServiceImpl implements BpmnSyncService {
   private Resource[] ensurePropertiesProvided() {
     Assert.notNull(bpmnSyncProperties.getDeploymentName(), "Application name must be provided.");
     Assert.notNull(bpmnSyncProperties.getBpmnVersion(), "BPMN version must be provided.");
-    Resource[] deployableResources = ResourceUtil.getDeployableResources();
-    Assert.notEmpty(deployableResources,
-        "No deployable resources found in classpath:bpmn or classpath:dmn folders.");
+    Resource[] deployableResources =
+        ResourceUtil.getDeployableResources(bpmnSyncProperties.getResourceLocation());
+    Assert.notEmpty(deployableResources, "No deployable *.bpmn or *.dmn files found under "
+        + bpmnSyncProperties.getResourceLocation() + ".");
     return deployableResources;
   }